+++ packages/blockchain-link-utils/src/tests/fixtures/blockfrost.ts
@@ -142,7 +142,7 @@ export default {
                 {
                     type: 'BLOCKFROST',
                     name: 'asset1hwnpal5vap799t6kkjmjf6myhse4zl2vu4ahzz',
-                    address:
+                    contract:
                         'b863bc7369f46136ac1048adb2fa7dae3af944c3bbb2be2f216a8d4f42657272795361707068697265',
                     symbol: 'BerrySapphire',
                     balance: '10',
@@ -151,7 +151,7 @@ export default {
                 {
                     type: 'BLOCKFROST',
                     name: 'asset1zvclg2cvj4e5jfz5vswf3sx0lasy79xn8cdap9',
-                    address: '02477d7c23b4c2834b0be8ca8578dde47af0cc82a964688f6fc95a7a47524943',
+                    contract: '02477d7c23b4c2834b0be8ca8578dde47af0cc82a964688f6fc95a7a47524943',
                     symbol: 'GRIC',
                     balance: '1',
                     decimals: 0,
@@ -490,7 +490,7 @@ export default {
                     {
                         type: 'BLOCKFROST',
                         name: 'asset1eevmdlaz5424s3663ypw8w4vyxdlxkm3lefz06',
-                        address:
+                        contract:
                             '2f712364ec46f0cf707d412106ce71ef3370f76e27fb56b6bb14708776657465726e696b4e657a6a6564656e79',
                         symbol: 'veternikNezjedeny',
                         balance: '1',
@@ -785,7 +785,7 @@ export default {
                         type: 'recv',
                         name: 'asset1eevmdlaz5424s3663ypw8w4vyxdlxkm3lefz06',
                         symbol: 'veternikNezjedeny',
-                        address:
+                        contract:
                             '2f712364ec46f0cf707d412106ce71ef3370f76e27fb56b6bb14708776657465726e696b4e657a6a6564656e79',
                         decimals: 0,
                         amount: '1',