+++ packages/suite/src/components/suite/modals/TransactionDetail/components/AmountDetails.tsx
@@ -221,7 +221,7 @@ export const AmountDetails = ({ tx, isTestnet }: AmountDetailsProps) => {
                             <FiatValue
                                 amount={formatAmount(transfer.amount, transfer.decimals)}
                                 symbol={transfer.symbol}
-                                tokenAddress={transfer.address}
+                                tokenAddress={transfer.contract}
                             />
                         }
                     />